+++ R/check.R
@@ -11,27 +11,24 @@
 #' @export
 check = function(server = NULL, token = NULL,
                  dir = ".", file = "config-uat.yml") {
-  summarise_setup(server, token)
-  user_details = summarise_user(get_server(), get_token())
+  check_list = list()
+  check_list$setup = summarise_setup(server, token)
+  check_list$user_details = summarise_user(get_server(), get_token())
+  check_list$versions = summarise_versions(get_server(), get_token())
+  check_list$sys_deps = check_sys_deps()
+  register_uat_user(get_server(), get_token(), account = check_list$user_details$username)
 
-  check_server_headers(get_server())
-
-  check_server_version(get_server(), get_token())
-
-  summarise_versions(get_server(), get_token())
-
-  register_uat_user(get_server(), get_token(), account = user_details$username)
-
-  cli::cli_h2("Starting checks")
+  cli::cli_h2("Starting deployment checks")
   r6_inits = init_r6_checks(dir = dir, file = file)
   lapply(r6_inits, function(r6) r6$check())
-  results = purrr::map_dfr(r6_inits, ~.x$get_log())
-  dplyr::arrange(results, .data$group, .data$short)
+  check_list$deployments = purrr::map_dfr(r6_inits, ~.x$get_log())
+  check_list$deployments = dplyr::arrange(check_list$deployments, .data$group, .data$short)
+  check_list
 }
 
 init_r6_checks = function(dir, file) {
   exports = getNamespaceExports("jrHealthCheckConnect")
-  check_exports = sort(exports[stringr::str_starts(exports, "check_")])
+  check_exports = sort(exports[stringr::str_starts(exports, "check_deploy")])
   r6_inits = lapply(check_exports, init_r6_check, dir = dir, file = file)
   r6_inits
 }