+++ R/set_key.R
@@ -14,9 +14,21 @@ get_server = function(clean = FALSE) {
   return(server)
 }
 
-get_token = function() .connect[["connect_api_key"]]
+get_token = function() {
+  token = .connect[["connect_api_key"]]
+  if (is.null(token)) {
+    token = get_value("CONNECT_API_KEY")
+    set_key("connect_api_key", token)
+  }
+  if (is.na(token)) stop("Can't find CONNECT_API_KEY")
+  return(token)
+}
 
-get_username = function() .connect[["username"]]
+get_username = function() {
+  user = .connect[["username"]]
+  if (is.null(user)) stop("Can't find username")
+  return(user)
+}
 
 get_value = function(env_name, value = NULL) {
   if (is.null(value)) value = Sys.getenv(env_name, NA)