+++ tests/testthat/test-check.R
@@ -1,7 +1,14 @@
 test_that("High level test", {
+  # Skip the deployment tests on the CI
+  # Takes a while
+  create_config(default = !isTRUE(as.logical(Sys.getenv("CI"))))
+
   rtn = check()
-  expect_s3_class(rtn, "tbl_df")
-  expect_gte(nrow(rtn), 3)
-  expect_equal(ncol(rtn), 5)
+
+  expect_true(is.list(rtn))
+  check_names = c("setup", "user_details", "versions", "sys_deps", "deployments") %in%  names(rtn)
+  expect_true(all(check_names))
+
+  expect_equal(ncol(rtn$deployments), 5)
   expect_error(check(server = "aaa.bbb"))
 })